Fixing navigation and adding new project generator page
